pyproject.toml
[project]
name = "chatserver"
version = "0.1.0"
description = "FastAPI backend for ChatLLM application with Gemini AI integration"
authors = [
    {name = "ChatLLM Team", email = "team@chatllm.local"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.0.0",
    "google-cloud-aiplatform>=1.38.0",
    "google-cloud-storage>=2.10.0",
    "google-cloud-documentai>=2.20.0",
    "firebase-admin>=6.5.0",
    "openai>=1.3.0",
    "websockets>=12.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "langchain>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-google-vertexai>=2.0.25",
    "langchain-community>=0.3.25",
    "langgraph>=0.4.8",
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.0",
    "pyjwt>=2.10.1",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.10.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.0",
]
knowledge = [
    "google-cloud-documentai>=2.20.1",
    "google-cloud-storage>=2.10.0",
    "PyPDF2>=3.0.1",
    "python-docx>=0.8.11",
]
full = [
    "google-cloud-documentai>=2.20.1",
    "google-cloud-storage>=2.10.0",
    "PyPDF2>=3.0.1",
    "python-docx>=0.8.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.10.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.0",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "google.generativeai.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=app",
    "--cov-report=html",
    "--cov-report=term-missing",
]
asyncio_mode = "auto"
